backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django~=4.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
flower==2.0.1
fuzzywuzzy==0.18.0
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.4
kombu==5.4.2
Levenshtein==0.26.0
MarkupSafe==2.1.5
msgpack==1.1.0
numpy~=2.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
prometheus_client==0.21.0
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.0
pytz==2024.2
RapidFuzz==3.10.0
redis==5.1.0
rpy2==3.5.16
service-identity==24.2.0
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.1.1
